Created jobs and flows folders, and separated job and flow files for standard ReactCA simulation and Bayesian Optimization simulation.
