+++ src/rxn_ca/workflow/__init__.py
@@ -9,16 +9,15 @@ Requires the [workflow] optional dependency:
 
 from .schemas import SimulationOutput, ReactionLibraryData
 from .jobs import setup_reaction_library, run_simulation, init_bo_campaign, bo_trial_step
-from .flows import create_simulation_flow
-from .bayesian_flow_makers import BOFlowMaker
+from .flows import create_simulation_flow, BOFlowMaker
 
 __all__ = [
     "SimulationOutput",
     "ReactionLibraryData",
     "setup_reaction_library",
     "run_simulation",
-    "create_simulation_flow",
     "init_bo_campaign",
     "bo_trial_step",
+    "create_simulation_flow",
     "BOFlowMaker",
 ]
