+++ src/rxn_ca/workflow/flows/__init__.py
@@ -1,0 +1,8 @@
+from .core import create_simulation_flow, create_multi_simulation_flow
+from .bayesian import BOFlowMaker
+
+__all__ = [
+    "create_simulation_flow",
+    "create_multi_simulation_flow",
+    "BOFlowMaker",
+]
