+++ src/rxn_ca/workflow/jobs/__init__.py
@@ -1,0 +1,9 @@
+from .core import setup_reaction_library, run_simulation
+from .bayesian import init_bo_campaign, bo_trial_step
+
+__all__ = [
+    "setup_reaction_library",
+    "run_simulation",
+    "init_bo_campaign",
+    "bo_trial_step",
+]
